requirements.txt
pandas
matplotlib
seaborn

psycopg2-binary
python-dotenv
flask
flask_cors
flask_migrate
flask_sqlalchemy

pytest

